Create CitiesOfWorld.csv
The file to be used as knowledge base for the AI agent to refine the location preferences for the job location. 

e.g. If the current user is residing in Portsmouth, and the agent is looking for the profiles in and around London and the agent have to shortlist 50 positions for the interviews, and only have 35 London based profiles, the agent has to extend the search not just limiting to London but additional UK geographies after exhausting the number of shortlists in London. The agent can use this file for identifying the cities, capitals, and the closest radius during the screening and profile ranking to source the additional profiles.